# Make fast_global_inits work again