--- a/src/analyses/base.ml
+++ b/src/analyses/base.ml
@@ -1620,7 +1620,7 @@ struct
   (**************************************************************************
    * Simple defs for the transfer functions
    **************************************************************************)
-  let zeroinit ctx v =
+  let zero_init ctx v =
     set ctx.ask ctx.global ctx.local (AD.from_var v) (zero_init_value v.vtype)
 
   let assign ctx (lval:lval) (rval:exp):store  =