By default, we generate `n*m` assign edges for a global array of dimensions `array[n][m]` when dealing with global initializers.

As an improvement, @vogler added the option `exp.fast_global_inits` where, for zero-initialized global arrays, only one assign to zero is generated: `array[0][0] = 0;`. This only works for the flat array domain, so we had to disable it when using the partitioned arrays (which we also do in SVCOMP).

**This PR makes this option compatible with partitioned arrays. To this end, we**

- Introduce `Base.zero_init_value t` to make a zero-initialized abstract value of type `t`.
   -  (In this step, we also remove unused arguments from `Base.init_value`, `Base.top_value`, and `Base.bot_value`)
- Add a new transfer function `zeroinit` that takes a `varinfo`
- Make `MyCFG.getGlobalInits file` return a list of `varinfo` to be zero-initialized on top of a list of edges for initialization
- Make `exp.fast_global_inits` the default and remove `--disable exp.fast_global_inits` from all test explicitly setting it

We chose to still generate the `array[0][0] = 0` edge, even when `exp.fast_global_inits` is enabled, causing `zeroinit array` to be called later.
This is so that it is safe to do nothing on `zeroinit` in all analyses except `base`. They still get the same assignment they got when `fast_global_inits` was implemented in the old way, so if they were sound before this change to `fast_global_inits`, they are still.

**The performance impact is quite big for large arrays:**

```
int global_array[50][200][20];

int main(void) {
  for(int i =0; i < 50; i++) {
      assert(global_array[i][42][7] == 0);
  }
}
```
Without `fast_global_inits`:
```
TOTAL                          157.453 s
  analysis                       157.450 s
    solving                         0.008 s

Memory statistics: total=293813.44MB, max=166.58MB, minor=293798.83MB, major=806.48MB, promoted=791.88MB
    minor collections=140102  major collections=24 compactions=0
```

With `fast_global_inits`:
```
TOTAL                           0.186 s
  analysis                        0.185 s
    solving                         0.009 s

Memory statistics: total=235.24MB, max=4.39MB, minor=233.53MB, major=3.47MB, promoted=1.76MB
    minor collections=112  major collections=3 compactions=0
```

Closes #122