@vogler 2020-11-09
Seems a bit heavy to introduce a transfer function for it.
Maybe use a special `zero_init` expression that `assign` can react to? After all, it's only a special case of `assign`.