@michael-schwarz 2020-11-09
> Maybe use a special `zero_init` expression that `assign` can react to? After all, it's only a special case of `assign`.

I also thought about that. My reasoning against it was that it introduces a new variant for CIL expression that we then need to handle everywhere we do anything with expressions (also inside CIL), and that we would have to modify the assign in every analysis.

The other thing I considered was adding some sort of boolean `zero_init` to the `assign` edge, but that also didn't look particularly nice.