@vogler 2020-11-09
I meant using something like `MyCFG.unknown_exp`, just as an int of the right type.
Maybe `MyCFG.all_indices` which is meant to abstract paths for all definite indices for the size of the type.
Every place that does not handle it will just fall back to top.