@michael-schwarz 2020-11-11
> I meant using something like `MyCFG.unknown_exp`, just as an int of the right type.

I tried this in #133. But I am not convinced it is nicer than this one here.