@michael-schwarz 2020-11-12
We fixed this the hacky way by mergingn #133 for now, when we want to support initializers properly (i.e. not just zero initialization), the way to go would probably be this one.